+++ proxy/src/main/java/com/velocitypowered/proxy/config/VelocityConfiguration.java
@@ -31,7 +31,7 @@ import com.velocitypowered.proxy.config.migration.ForwardingMigration;
 import com.velocitypowered.proxy.config.migration.KeyAuthenticationMigration;
 import com.velocitypowered.proxy.config.migration.MotdMigration;
 import com.velocitypowered.proxy.util.AddressUtil;
-import com.velocitypowered.proxy.util.TFOUtil;
+import com.velocitypowered.proxy.util.TcpFastOpenUtil;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.io.IOException;
 import java.net.InetSocketAddress;
@@ -360,11 +360,11 @@ public class VelocityConfiguration implements ProxyConfig {
   }
 
   public boolean useTcpFastOpenClient() {
-    return advanced.isTcpFastOpen() && TFOUtil.IS_SUPPORTING_TCP_FASTOPEN_CLIENT;
+    return advanced.isTcpFastOpen() && TcpFastOpenUtil.IS_SUPPORTING_TCP_FASTOPEN_CLIENT;
   }
 
   public boolean useTcpFastOpenServer() {
-    return advanced.isTcpFastOpen() && TFOUtil.IS_SUPPORTING_TCP_FASTOPEN_SERVER;
+    return advanced.isTcpFastOpen() && TcpFastOpenUtil.IS_SUPPORTING_TCP_FASTOPEN_SERVER;
   }
 
   public Metrics getMetrics() {
